Fixed eval bug
Bug was a problem with the chaining offset.
In creating the offset, the second input block was not grabbing the
correct block. It grabbed the outputLabel when it should have grabbed
the inputLabel. And it was grabbing the ith label when it should
have grabbed the (2*i)th label (because there are twice as many
input labels.

Also moved the garbler and evaluation (formerly src/main.c:go2()) test
into src/tests.c.