Adds comm. of inputmapping/instrs. Adding network
In processing of writing garbler_go() and evaluator_go(),
the core fucntions of online communication and processing by
both parties. Added some functions so that inputmapping and instructions
can be serialized into a buffer and sent over the network.

Next step is doing OT between the parties.

I've been hardcoding somethings related to 22Adder complex circuit,
will need to go back and abstract thigns out.

Also, added a sleep in evaluator_go() because running too fast to make
connection. The problem is that there are two connections. The first one
is to communicate the chained garbled circuits (offline) and the second
one is the online stuff.