debugging outputmap error.
So mapoutputs, the final step of evaluation,
is not working properly. It seems to be a problem with
MixColumns in particular. Trying to figure out what.
At this point, it seems like it may be an error with
JustGarble. Could also be my usage of MixColumns in
buildAESRoundComponent.

Checked doing aes round component wihtout mixcolumns,
and things seem to work fine. Trying other example.
Should try a circuit that *only* mixes columns. See if that works,
and debug from there.