added timings for aes
Sometimes stops working because the garbler and evaluator network
connection isn't working. What is happening is that the evaluator is
moving too fast, so the garbler hasn't yet opened up for a connection.
The solution is to slow down the evaluator code, which can be done by
uncommenting the line in `test/2pc_aes.c:aes_eval_on` `sleep(1)`.