modified makefile to compile with new JG.
Makefile still a little shaky because it doesn't use JG's makefile, but
will compile JG stuff if it needs to.

There are also a few things in JG which are changed on my local version,
but not pushed. They are `static inline` not compiling and
blocksEqual needs to be in public API.